lapr4/1181498-lei_isep_2019_20_sem4_2dd_1181498_1181500_1181255_1181529-3197f42c0ac1/eapli.base/base.core/src/test/java/eapli/base/product/domain/ProductDescriptionTest.java
package eapli.base.product.domain;

import static org.junit.Assert.*;
import org.junit.Before;
import org.junit.Test;


public class ProductDescriptionTest {

    ProductDescription instance;

    @Before
    public void before() {
        instance = new ProductDescription("Blocos","Blocos Borracha");
    }

    /**
     * Test of GetBriefDescription method, of class ProductDescription.
     */
    @Test
    public void testGetBriefDescription() {
        instance = new ProductDescription("Blocos","Blocos Borracha");
        String exp = "Blocos";
        assertEquals(exp,instance.getBriefDescription());
    }



    /**
     * Test of GetCompleteDescription method, of class ProductDescription.
     */
    @Test
    public void testGetCompleteDescription() {
        instance = new ProductDescription("Blocos","Blocos Borracha");
        String exp = "Blocos Borracha";
        assertEquals(exp,instance.getCompleteDescription());
    }
}